% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getauthorrecordraw.R
\name{getauthorrecordraw}
\alias{getauthorrecordraw}
\title{Provides extended info about author}
\usage{
getauthorrecordraw(id, code = NA)
}
\arguments{
\item{id}{Short-ID.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
data.frame. Note: columns \code{article} and \code{paper} are data.frames.
}
\description{
Provides extended info about author
}
\examples{
\dontrun{
getauthorrecordraw('pmo915', code = YOURCODE)
}

}
