% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getauthorsforitem.R
\name{getauthorsforitem}
\alias{getauthorsforitem}
\alias{get_authors_for_item}
\title{Get Authors for an item}
\usage{
getauthorsforitem(x, code = NA)

get_authors_for_item(x, code = NA)
}
\arguments{
\item{x}{RePEc handle for an item.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
data.frame.
}
\description{
Get Authors for an item
}
\examples{
\dontrun{
getauthorsforitem('RePEc:eee:enepol:v:78:y:2015:i:c:p:41-50', code = YOURCODE)
}

}
