% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getauthorshortid.R
\name{getauthorshortid}
\alias{getauthorshortid}
\title{Get Short-ID and name from string in name or email}
\usage{
getauthorshortid(x, code = NA)
}
\arguments{
\item{x}{string.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
data.frame
}
\description{
Get Short-ID and name from string in name or email
}
\examples{
\dontrun{
getauthorshortid('christian', code = YOURCODE)
}

}
