% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getauthorstats.R
\name{getauthorstats}
\alias{getauthorstats}
\alias{get_author_stats}
\title{Provides Author's statistics}
\usage{
getauthorstats(id, code = NA)

get_author_stats(id, code = NA)
}
\arguments{
\item{id}{Short-ID.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
Author's statistics.
}
\description{
Provides Author's statistics
}
\examples{
\dontrun{
getauthorstats('pmo915', code = YOURCODE)
}

}
