% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getauthortwitter.R
\name{getauthortwitter}
\alias{getauthortwitter}
\alias{get_author_twitter}
\title{Provides Twitter handle}
\usage{
getauthortwitter(id, code = NA)

get_author_twitter(id, code = NA)
}
\arguments{
\item{id}{Short-ID.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
String with the Twitter author linked to the Short-ID.
}
\description{
Provides Twitter handle
}
\examples{
\dontrun{
getauthortwitter('pmo915', code = YOURCODE)
}

}
