% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgenealogy.R
\name{getgenealogy}
\alias{getgenealogy}
\alias{get_genealogy}
\title{Provides Author's genealogy}
\usage{
getgenealogy(id, code = NA)

get_genealogy(id, code = NA)
}
\arguments{
\item{id}{Short-ID.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
Author's genealogy.
}
\description{
Provides Author's genealogy
}
\examples{
\dontrun{
getgenealogy('pmo915', code = YOURCODE)
}

}
