% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethindex.R
\name{gethindex}
\alias{gethindex}
\title{Provides H-index}
\usage{
gethindex(id, code = NA)
}
\arguments{
\item{id}{Short-ID.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
H-index.
}
\description{
Provides H-index
}
\examples{
\dontrun{
gethindex('pmo915', code = YOURCODE)
}

}
