% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinstrecord.R
\name{getinstrecord}
\alias{getinstrecord}
\alias{get_inst_record}
\title{Get record of Institution}
\usage{
getinstrecord(x, code = NA)

get_inst_record(x, code = NA)
}
\arguments{
\item{x}{RePEc handle for an institution}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
data.frame.
}
\description{
Get record of Institution
}
\examples{
\dontrun{
getinstrecord('RePEc:edi:dechiit', code = YOURCODE)
}

}
