% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getjelforitem.R
\name{getjelforitem}
\alias{getjelforitem}
\alias{get_jel_foritem}
\title{Get JEL codes for item}
\usage{
getjelforitem(id, code = NA)

get_jel_foritem(id, code = NA)
}
\arguments{
\item{id}{RePEc handle for an item.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
Vector with JEL codes.
}
\description{
Get JEL codes for item
}
\examples{
\dontrun{
getjelforitem('RePEc:eee:enepol:v:78:y:2015:i:c:p:41-50', code = YOURCODE)
}

}
