% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repec.R
\docType{package}
\name{repec-package}
\alias{repec}
\alias{repec-package}
\title{repec: access RePEc data through API}
\description{
\code{repec} provides a way to interact with the RePEc API (Application
Programming Interface) in order to get various information stored
on RePEc about authors, items, institutions, etc.
}
\details{
By using the functions of this package, you will be able to import
RePEc information in R. You will need to ask for an access code
by following the instructions at https://ideas.repec.org/api.html
The code will be needed when calling most of the functions by using
the \code{code} parameter. One way to avoid passing explicitly the code
is to create the \code{REPEC_API_KEY} environment variable. The code below
can be put in your \code{.RProfile} so that it is always available when
you start R:\preformatted{# YOURCODE is your personal access code
Sys.setenv(REPEC_API_KEY = 'YOURCODE')
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chrMongeau/repec}
  \item Report bugs at \url{https://github.com/chrMongeau/repec/issues}
}

}
\author{
\strong{Maintainer}: Christian Alexander Mongeau Ospina \email{christian+rpkg-repec@mongeau.net} (0000-0002-4047-3924)

}
