% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repec_api_with_id.R
\name{repec_api_with_id}
\alias{repec_api_with_id}
\title{Generic API call, with an ID}
\usage{
repec_api_with_id(id = NA, method = NA, code = NA)
}
\arguments{
\item{id}{Short-ID.}

\item{method}{API method to call.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
String with the Twitter author linked to the Short-ID.
}
\description{
Generic API call, with an ID
}
\examples{
\dontrun{
repec_api_with_id(id = 'pmo915',
  method = 'getauthortwitter', code = YOURCODE)
}

}
