% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areyouthere.R
\name{areyouthere}
\alias{areyouthere}
\title{Verifies whether the server is alive}
\usage{
areyouthere()
}
\value{
TRUE if the server if alive, FALSE otherwise.
}
\description{
Verifies whether the server is alive
}
\examples{
\dontrun{
areyouthere()
}

}
