% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getauthornep.R
\name{getauthornep}
\alias{getauthornep}
\title{Get the NEP fields for the papers of an author}
\usage{
getauthornep(x, code = NA)
}
\arguments{
\item{x}{Short-ID.}

\item{code}{User code required to use API. Included in every API call.}
}
\value{
data.frame
}
\description{
Get the NEP fields for the papers of an author
}
\examples{
\dontrun{
getauthornep('pmo915', code = YOURCODE)
}

}
