% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getauthorrecordfull.R
\name{getauthorrecordfull}
\alias{getauthorrecordfull}
\title{Provides info about author, mostly in the form of handles}
\usage{
getauthorrecordfull(x, code = NA)
}
\arguments{
\item{x}{Short-ID.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
data.frame
}
\description{
Provides info about author, mostly in the form of handles
}
\examples{
\dontrun{
getauthorrecordfull('pmo915', code = YOURCODE)
}

}
