% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfirstpubyear.R
\name{getfirstpubyear}
\alias{getfirstpubyear}
\alias{get_first_pub_year}
\title{Provides Author's first publication year}
\usage{
getfirstpubyear(id, code = NA)

get_first_pub_year(id, code = NA)
}
\arguments{
\item{id}{Short-ID.}

\item{code}{User code required to use API. Included in every API call.
If not supplied, the REPEC_API_KEY environment variable will be used.}
}
\value{
Author's first publication year.
}
\description{
Provides Author's first publication year
}
\examples{
\dontrun{
getfirstpubyear('pmo915', code = YOURCODE)
}

}
