% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatismyip.R
\name{whatismyip}
\alias{whatismyip}
\title{Returns the IP address you use to call the API}
\usage{
whatismyip()
}
\value{
String with IP.
}
\description{
Returns the IP address you use to call the API
}
\examples{
\dontrun{
whatismyip()
}

}
